+++ packages/components/atoms/f-button/CHANGELOG.md
@@ -6,7 +6,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 v3.1.0
 ------------------------------
-*October 18, 2021*
+*October 20, 2021*
 
 ### Changed
 - Add spacing to buttons following a paragraph tag. This matches the current behaviour in Fozzie.
@@ -14,6 +14,7 @@ v3.1.0
 - `f-vue-icons` package version bump.
 - Button story now shows all the types and sizes of the button component (primary, secondary, outline and ghost).
 - Changed knobs to controls for switching button props.
+- Separate Demo story to still show one button in isolation.
 - Separate story for Link Button to be able to run tests properly.
 - Updated the tests in line with the changes.